fix: Handle tunnel reqs/resps using channels
Previously responses were being incorrectly read causing data corruption because
multiple request goroutines were unsafely reading from the same connection.

The new implementation created a channel to forward (tunnel) requests to, which would then
send it to the mmar client connection, process them, then return response to another
channel that belongs to the request's goroutine.

This ensures no request is reading a response
that doesn't belong to it. This may potentially cause a bottleneck, will look into resolving it
if that becomes an issue.