feat: Add heartbeat from server->client
This heartbeat helps determine if mmar client has disconnected from server. It only runs
if no data has been exchanged between the mmar server and client for some interval of time.
If it successfully sends the heartbeat, that means the client is still connected, otherwise
it has been disconnected, and its safe to stop the goroutines related to that Tunnel.