fix: Properly cancel client ProcessTunnelMessages
This fixes a bug in the select that checks if the context
is cancelled to kill the ProcessTunnelMessages goroutine.
Initially it would enter the default case and block indefinitely
preventing the ctx.Done() case from ever reaching.

To fix this, we add a read deadline to allow for the blocking to
end and loop through again, giving a chance to check the cancel
condition.