fix: Do not block reqChan until req serialized
Previously we handled serializing requests inside the incoming
requests channel loop. Which meant that requests that take long
to serialize and eventually timeout because they misconfigured
would block other requests until the timeout occurs, which is
not great. The new implementation now handles serializing the
request before it is sent to the incoming request channel to be
tunnelled, allowing for other requests to flow through freely.